[Feat] 선택한 스테이지 정보 저장
선택한 스테이지의 정보를 저장하여 그 정보를 받아 씬 전환에 사용하도록 함
