+++ CrazyArcade/Assets/Scripts/Item/ItemData/Needle.asset
@@ -14,3 +14,4 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   Id: 0
   Icon: {fileID: 21300000, guid: 3f7be2185b5c3ae40a8044cf8f53a6c8, type: 3}
+  Condition: 117565546
