+++ CrazyArcade/Assets/Scripts/UI/Lobby.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f10820f48f3c15749afa2ce1a9973bcc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
