+++ CrazyArcade/Assets/Scripts/UI/Lobby/GameStartButton.cs
@@ -1,0 +1,21 @@
+using UnityEngine;
+using UnityEngine.SceneManagement;
+using UnityEngine.UI;
+
+public class GameStartButton : MonoBehaviour
+{
+    private Button _button;
+
+    private void Awake()
+    {
+        _button = GetComponent<Button>();
+        _button.onClick.AddListener(
+            LoadStage
+         );
+    }
+
+    public void LoadStage()
+    {
+        SceneManager.LoadScene(GameManager.Instance.SelectedStage);
+    }
+}
