+++ CrazyArcade/Assets/Scripts/UI/Lobby/MapSelection/MapData.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+[CreateAssetMenu]
+public class MapData : ScriptableObject
+{
+    public int StageNumber;
+}
