+++ CrazyArcade/Assets/Scripts/UI/Lobby/MapSelection/MapData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bc8cfb80b2ed79541af0fd9becfed966
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
