+++ CrazyArcade/Assets/Scripts/UI/Lobby/MapSelection/MapSelectionButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b2a79465351917c4fb1acf264e7337a9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
