+++ CrazyArcade/Assets/Scripts/UI/Lobby/MapSelection/Stage1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f771b71d23a9fa54c869fe11b6dda09d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
