+++ CrazyArcade/Assets/Scripts/UI/Lobby/MapSelection/Stage2.asset
@@ -1,0 +1,15 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: bc8cfb80b2ed79541af0fd9becfed966, type: 3}
+  m_Name: Stage2
+  m_EditorClassIdentifier: 
+  StageNumber: 2
