+++ CrazyArcade/Assets/Scripts/UI/Lobby/MapSelection/Stage2.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 97bc3ee2138c29746863b5b3cf596c52
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
