--- CrazyArcade/Assets/Scripts/UI/LobbyButton.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f10820f48f3c15749afa2ce1a9973bcc
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
