--- CrazyArcade/Assets/Scripts/UI/LobbyButton/GameStartButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c0bee5ba071061b469d8d1f61cf0ae3e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
