+++ CrazyArcade/Assets/Scripts/UI/Stage/OutButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: acf25c8f19ffd43458cfd15a34d22638
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
