--- CrazyArcade/Assets/Scripts/UI/StageButton.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 813b2858e889fd149888ffd00f826471
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
