--- CrazyArcade/Assets/Scripts/UI/StageButton/OutButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: acf25c8f19ffd43458cfd15a34d22638
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
