+++ CrazyArcade/Assets/TextMesh Pro/Fonts/1_SeoulNamsanM.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cf3bcad1a631d674883f09be48bba880
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
