+++ CrazyArcade/Assets/TextMesh Pro/Fonts/DNFBitBitOTF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9c5c089cdb5402b40ab791d641a4822b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
