+++ CrazyArcade/Assets/TextMesh Pro/Fonts/Daum_Regular SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 907a87fec52731b41969ce8768e9db5c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
