+++ CrazyArcade/Assets/TextMesh Pro/Fonts/Daum_Regular.ttf.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: d7abdb2ea298fec4e847c6ce11310a39
+TrueTypeFontImporter:
+  externalObjects: {}
+  serializedVersion: 4
+  fontSize: 16
+  forceTextureCase: -2
+  characterSpacing: 0
+  characterPadding: 1
+  includeFontData: 1
+  fontNames:
+  - Daum
+  fallbackFontReferences:
+  - {fileID: 12800000, guid: 13f952b6c3c5ab745af1af9406ccfe41, type: 3}
+  customCharacters: 
+  fontRenderingMode: 0
+  ascentCalculationMode: 1
+  useLegacyBoundsCalculation: 0
+  shouldRoundAdvanceValue: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
