+++ CrazyArcade/Assets/TextMesh Pro/Fonts/SeoulNamsanM.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2e4d52cd1a3b81e459c0ee41d5c1469f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
